requirements.txt
python-dotenv
fastapi
uvicorn
celery
redis
google-generativeai